Refactor semantic element logical positions
It's still not totally clear to me how semantic element tags and
Bootstrap required structure are intended to play nice with
each other.  What I've tried for here is at least when
hovering over the element name in the dev tools it highlights
the actual section the semantic element refers to.  This wasn't
intuitive at first.